scraper/requirements.txt
mysql_connector_repackaged==0.3.1
selenium==4.31.0
webdriver_manager==4.0.2
mysql-connector==2.2.9